// apps/jobs/web/src/lib/api/generated/model/application.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire Jobs API
 * Jobs mini-app API
 * OpenAPI spec version: 0.1.0
 */
import type { JobRef } from './jobRef';
import type { ApplicationApplicant } from './applicationApplicant';
import type { ApplicationAnswersItem } from './applicationAnswersItem';
import type { ApplicationStatus } from './applicationStatus';
import type { ApplicationStatusHistoryItem } from './applicationStatusHistoryItem';

export interface Application {
  id: string;
  job: JobRef;
  applicant: ApplicationApplicant;
  /** @nullable */
  headline: string | null;
  /** @nullable */
  coverNote: string | null;
  answers: ApplicationAnswersItem[];
  /** @nullable */
  portfolioUrl: string | null;
  status: ApplicationStatus;
  statusHistory: ApplicationStatusHistoryItem[];
  /** @nullable */
  recruiterNotes: string | null;
  createdAt: string;
}